% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_variables}
\alias{add_variables}
\title{Check a data frame for the existence of columns. If not present, create variables with a default value.}
\usage{
add_variables(df, varnames, default = NA)
}
\arguments{
\item{df}{Data frame that is checked (and optionally expanded with new variables)}

\item{varnames}{The variable names that are searched in the df}

\item{default}{The default value for newly created columns}
}
\description{
Check a data frame for the existence of columns. If not present, create variables with a default value.
}
