% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_barchart.R
\name{BC_h}
\alias{BC_h}
\title{BC_h}
\usage{
BC_h(cat_labels, values, colors, title = "", rev = FALSE)
}
\arguments{
\item{cat_labels}{a character vector of category labels for the bars.}

\item{values}{a numeric vector of values for each category.}

\item{colors}{a character vector of colors for each category.}

\item{title}{The title of the plot.}

\item{rev}{Should the order of categories be reversed?}
}
\value{
a plotly object of the stacked bar chart.
}
\description{
This function creates a horizontal stacked bar chart in plotly with bar labels displaying category labels, values, and percentages.
}
\examples{
BC_h(
  cat_labels=c("Yes", "No", "not applicable"),
  values=c(10, 2, 4),
  colors=c("#90c916", "#FED976", "#eeeeee"),
  title="The publication contained a preregistered replication attempt")

BC_h(
  cat_labels=c("Yes", "No"),
  values=c(10, 2),
  colors=c("#90c916", "#FED976"))

BC_h(cat_labels=c("Yes"), values=c(10), colors=c("#90c916"))

}
