% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkcharts.R
\name{circle_layer_html}
\alias{circle_layer_html}
\title{Four layered circles}
\usage{
circle_layer_html(outer_width = 60, colors)
}
\arguments{
\item{outer_width}{Circle radius in pixels.}

\item{colors}{A vector of four colors, from innermost to outermost color}
}
\description{
TODO: Provide color ramp breakpoints as parameters
}
