% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_RRS.R
\name{compute_RRS}
\alias{compute_RRS}
\title{Compute Relative Rigor Score (RRS)}
\usage{
compute_RRS(applicant)
}
\arguments{
\item{applicant}{The applicant data that has been preprocessed by the \code{read_RESQUE} function.}
}
\value{
A tibble containing the dimension (category), maximum points, and relative score for each category, suitable for creating a radar chart.
}
\description{
This function computes a relative rigor score based on the indicators provided from an applicant.
}
