% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit_barchart.R
\name{credit_barchart}
\alias{credit_barchart}
\title{Display a barchart with CRediT roles}
\usage{
credit_barchart(credit_dat, ordered = FALSE, compact = FALSE)
}
\arguments{
\item{credit_dat}{The raw data in the format of the \code{applicant()$credit} field. This is a data frame with two relevant columns: \code{Role} (the CRediT role) and \code{Degree} (one of "Lead", "Equal", "Support", "NoRole", or "not applicable"). Roles can appear multiple times (e.g., for multiple papers) and will be aggregated in the plot.}

\item{ordered}{If \code{FALSE}, keep alphabetical order. If \code{TRUE}, sort by strongest contribution.}

\item{compact}{If \code{FALSE}, show all 14 roles. If \code{TRUE}, aggregate the 14 roles into 4 categories.}
}
\description{
Display a barchart with CRediT roles
}
\examples{
credit_dat <- data.frame(
  Role = rep(c("Conceptualization", "Data Curation", "Formal Analysis",
    "Funding Acquisition", "Investigation", "Methodology",
    "Project Administration", "Resources", "Software", "Supervision",
    "Validation", "Visualization", "Writing: Original draft",
    "Writing: Review & Editing"), 10),
  Degree = sample(c("Lead", "Equal", "Support", "NoRole", "not applicable"),
    prob = c(0.1, 0.2, 0.3, 0.3, 0.1), size=140, replace=TRUE)
)
credit_barchart(credit_dat)
credit_barchart(credit_dat, ordered=TRUE)
}
