% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{evaluate_condition_in_context}
\alias{evaluate_condition_in_context}
\title{Evaluate a condition in a context}
\usage{
evaluate_condition_in_context(condition, research_output)
}
\arguments{
\item{condition}{The logical conditions in the \code{scoring} section of core-pub.json. This can either be the \code{score} -> \code{not_applicable} condition or the \code{score} -> \code{condition} condition}

\item{research_output}{The research output to be scored. This is the latter part of the export json with the actual item values of one specific publication.}
}
\description{
Evaluate a condition in a context
}
