% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{extract_indicators}
\alias{extract_indicators}
\title{Extract Indicators from Applicant List}
\usage{
extract_indicators(applicant_list)
}
\arguments{
\item{applicant_list}{A list containing the RESQUE jsons from multiple applicants.}
}
\value{
The function returns a data frame where each row corresponds to one
publication and indicators are in columns.
}
\description{
This function processes a list of applicants, extracting specific indicators
from each applicant's data and compiling them into a single data frame. It
also includes metadata such as ORCID, LastName, and ExternalRaterName from
the applicant's profile.
}
