% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{extract_overall_scores}
\alias{extract_overall_scores}
\title{Extract RRS Overall Score for Each Applicant}
\usage{
extract_overall_scores(applicant_list)
}
\arguments{
\item{applicant_list}{A list containing the RESQUE jsons from multiple applicants.}
}
\value{
A data frame containing the RRS overall score for each applicant.
}
\description{
This function processes a list of applicants, extracting the Relative Rigor Score (RRS)
or each applicant and compiling them into a single data frame. It includes
metadata such as ORCID, LastName, and ExternalRaterName from the applicant's profile.
}
