% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_list.R
\name{format_names}
\alias{format_names}
\title{Format Author Names of a single OpenAlex work object}
\usage{
format_names(works, alphabetical = TRUE)
}
\arguments{
\item{works}{A data frame containing author information.}

\item{alphabetical}{Logical indicating whether to sort the names alphabetically by last name (default: TRUE).}
}
\value{
A character vector of formatted names.
}
\description{
Formats the names of authors in a data frame.
}
