% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkcharts.R
\name{get_color}
\alias{get_color}
\title{Convert RRS values to green-red color colding (vectorized)}
\usage{
get_color(values)
}
\arguments{
\item{values}{RRS values (ranging from 0 to 1)}
}
\description{
TODO: Provide breakpoints as parameters
}
