% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{get_json_object}
\alias{get_json_object}
\title{Get JSON Object}
\usage{
get_json_object(url)
}
\arguments{
\item{url}{The URL to make the GET request to.}
}
\value{
An R object representing the converted JSON response.
}
\description{
Make a GET request to a specified URL and convert the JSON response to an R object using the jsonlite package.
}
