% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_list.R
\name{get_reference_list}
\alias{get_reference_list}
\title{Convert OpenAlex works into formatted reference list (in HTML)}
\usage{
get_reference_list(
  works,
  alphabetical = TRUE,
  journal_name = FALSE,
  bullet = TRUE
)
}
\arguments{
\item{works}{A data frame containing information about the works to be referenced. Each row of the data frame represents a single work, and columns represent different attributes of the works.}

\item{alphabetical}{Logical value indicating whether the references should be sorted alphabetically by author's last name. Default is TRUE.}

\item{journal_name}{Logical value indicating whether the journal name should be displayed. Default is FALSE.}

\item{bullet}{Logical value indicating whether the references should be formatted as an unordered list (\if{html}{\out{<ul>}}) or separated by line breaks (\if{html}{\out{<br>}}). Default is TRUE.}
}
\value{
A character vector containing the formatted APA references.
}
\description{
This function generates a reference list for a given set of OpenAlex works. The formatting is inspired by APA style, but it does (by default) omit the journal name and can re-order authors alphabetically
}
