% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is.nulla}
\alias{is.nulla}
\title{Check if an object is NULL, NA, or length zero}
\usage{
is.nulla(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
A logical scalar: \code{TRUE} if \code{x} is considered "null-like",
otherwise \code{FALSE}.
}
\description{
This utility function returns \code{TRUE} if the input is \code{NULL},
has length zero, or is an \code{NA} of length one. Otherwise returns \code{FALSE}.
}
\examples{
is.nulla(NULL)
is.nulla(NA)
is.nulla(character(0))
is.nulla(1)   # FALSE

}
