% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkcharts.R
\name{ministack}
\alias{ministack}
\title{Create a mini horizontal stacked bar chart}
\usage{
ministack(
  value,
  breaks = c(0.15, 0.3),
  colors = c("red", "#f7cf07", "#3af72c")
)
}
\arguments{
\item{value}{numeric value to be represented on the gauge (must be between 0 and 1)}

\item{breaks}{numeric vector specifying the break points for coloring the gauge steps (must be two elements)}

\item{colors}{character vector specifying the colors for each step in the gauge (must be three elements)}
}
\value{
a plotly object representing the mini gauge plot
}
\description{
This function creates a mini horizontal stacked bar chart using the plotly package. The gauge represents a value within a range, divided into color-coded steps and highlighting a threshold value.
}
\examples{
ministack(0.25)
}
