% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_RESQUE.R
\name{read_RESQUE}
\alias{read_RESQUE}
\title{Read a single RESQUE JSON file and do basic preprocessing
(elaborated preprocessing happens in \code{preprocess()}).}
\usage{
read_RESQUE(file, verbose = FALSE)
}
\arguments{
\item{file}{Path to RESQUE JSON file}

\item{verbose}{Show diagnostic information?}
}
\description{
Read a single RESQUE JSON file and do basic preprocessing
(elaborated preprocessing happens in \code{preprocess()}).
}
