% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_RRS.R
\name{RRS_category_summary}
\alias{RRS_category_summary}
\title{Display information about the scoring categories}
\usage{
RRS_category_summary(applicant, selector = c("P_"))
}
\arguments{
\item{applicant}{The applicant data that has been imported by the \code{read_RESQUE} function.}

\item{selector}{The prefix(es) of indicators that should be selected with \code{starts_with()}}
}
\value{
Nothing (called for it's print output)
}
\description{
This function shows a table how many indicators are assigned to each scoring
category. This counts all indicators that are defined in the packs.json and
contribute to scoring (regardless of whether an indicator is missing in that
specific applicant).
}
