% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRS_radarchart.R
\name{RRS_radarchart}
\alias{RRS_radarchart}
\title{Radar Chart for Relative Rigor Score}
\usage{
RRS_radarchart(radar_dat, title = "", base_size = 14)
}
\arguments{
\item{base_size}{Base font size}

\item{radardat}{The output form the \code{compute_RRS} function.}
}
\value{
A radar chart visualizing the rigor profile for the applicant, showing the distribution of scores across different dimensions.
}
\description{
This function generates a radar chart that visualizes the rigor profile for a given applicant's data.
}
\examples{
radar_dat <- data.frame(
  dimension = c("Open Data", "Open Materials", "Preregistration", "Reproducible Code \n& Verification", "Theorizing \n& Formal Modeling"),
  max_points = rep(10, 5),
  rel_score=c(0.5, 0.3, 0.1, 0.6, 0.2),
  xstart = 0:4,
  xend = 1:5,
  xmid=0:4 + 0.5
)
RRS_radarchart(radar_dat)

}
