% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_all}
\alias{score_all}
\title{Score multiple research outputs}
\usage{
score_all(research_outputs, verbose = FALSE)
}
\arguments{
\item{research_outputs}{A list of research outputs to be scored}

\item{verbose}{A logical value indicating if verbose output should be printed}
}
\value{
A list with scores, number of scored research outputs, and overall score
}
\description{
Returns a list with the following elements:
\itemize{
\item scores: a list of scores for each research output
\item scored_research_outputs: the number of research outputs that were scored
\item overall_score: the average score of all scored research outputs. Each output is weighted equally.
}
}
