% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_all_from_file}
\alias{score_all_from_file}
\title{Score all research outputs from a file}
\usage{
score_all_from_file(path, verbose = FALSE)
}
\arguments{
\item{path}{The file path where the research outputs are stored.}

\item{verbose}{Logical value indicating whether to print additional information. Default is \code{FALSE}.}
}
\value{
A list with elements scores, scored_research_outputs, and overall_score.
}
\description{
Reads the research outputs from the specified file and scores them using \code{\link{score_all}}.
Returns a list with the following elements:
\describe{
\item{scores}{a list of scores for each research output}
\item{scored_research_outputs}{the number of research outputs that were scored}
\item{overall_score}{the average score of all scored research outputs}
}
}
\examples{
\dontrun{
score_all_from_file("research_outputs.json", verbose = TRUE)
}

}
