% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_forms.R
\name{test_forms_json_structure}
\alias{test_forms_json_structure}
\title{Test structure of a "forms" JSON file}
\usage{
test_forms_json_structure(path)
}
\arguments{
\item{path}{Character scalar. Path to the JSON file to be tested.}
}
\value{
This function is called for its side effect of running
\pkg{testthat} expectations. It returns the result of the
\code{testthat::test_that()} call (invisibly).
}
\description{
Runs a set of \pkg{testthat} checks to verify that a JSON file containing
form definitions conforms to the expected schema. These json files are used
to replace the metadata of an existing json with exported RESQUE data
(e.g. if the scoring rules have been updated).
}
\details{
The core checks performed are:
\itemize{
\item The file contains valid JSON.
\item The top-level JSON object contains only a single element named
\code{"forms"}.
\item The \code{"forms"} object contains at least the elements
\code{"meta"}, \code{"pub"}, and \code{"config"}.
\item The \code{"pub"} element contains non-empty \code{"versions"} and
\code{"scoring"} elements.
\item The top-level \code{"config"} element contains a non-empty
\code{"score_categories"} element.
}

In addition, the function performs several \strong{additional} schema checks:
\itemize{
\item The file path exists.
\item Core forms (\code{"meta"}, \code{"pub"}, \code{"data"},
\code{"software"}) share a common structure (e.g. must contain
\code{title}, \code{versions}, \code{date}, \code{scoring},
\code{config}, \code{elements}).
\item Every element within \code{elements} has at least
\code{"id"} and \code{"type"} fields.
\item Each entry in \code{score_categories} contains at least
a \code{"title"} field.
\item \code{forms$config} contains meta-entries for
\code{"meta"}, \code{"pubs"}, \code{"data"}, and \code{"software"}.
}

These additional checks are intended to catch common schema deviations and
can be reviewed, modified, or removed as needed.
}
\examples{
\dontrun{
  test_forms_json_structure("inst/extdata/v0.8.1forms.json")
}

}
