% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{unCamel}
\alias{unCamel}
\title{For all cells in selected columns, replace CamelCase with spaces between words}
\usage{
unCamel(df, cname)
}
\arguments{
\item{df}{data.frame}

\item{cname}{Column names in which cells should be uncameled.}
}
\description{
For all cells in selected columns, replace CamelCase with spaces between words
}
