% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{unCamel0}
\alias{unCamel0}
\title{Replaces CamelCase in a vector with spaces between words}
\usage{
unCamel0(x)
}
\arguments{
\item{x}{A character vector which should be uncameled.}
}
\description{
Replaces CamelCase in a vector with spaces between words
}
