% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_json.R
\name{validate_json}
\alias{validate_json}
\title{Fix some problems with old exported json files}
\usage{
validate_json(file, update_forms = FALSE, verbose = TRUE)
}
\arguments{
\item{file}{Path to RESQUE JSON file}

\item{update_forms}{Should the \code{forms} node in the json (which contains the scoring information) be overwritten with the current version?}

\item{verbose}{Show diagnostic information?}
}
\description{
Fix some problems with old exported json files
}
