% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_RESQUE.R
\name{read_RESQUE}
\alias{read_RESQUE}
\title{Read and preprocess a single RESQUE JSON file}
\usage{
read_RESQUE(file, verbose = FALSE)
}
\arguments{
\item{file}{Path to RESQUE JSON file}

\item{verbose}{Show diagnostic information?}
}
\description{
Read and preprocess a single RESQUE JSON file
}
