% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_overview.R
\name{render_overview}
\alias{render_overview}
\title{Render an HTML overview of multiple applicants}
\usage{
render_overview(
  json_folder,
  output_file = NA,
  template = NA,
  anonymous = FALSE,
  clear_cache = FALSE
)
}
\arguments{
\item{json_folder}{The path to a folder with multiple JSON files}

\item{output_file}{The file name (optionally including a path) of the output report. If NA, it uses the folder name plus the current date-time as filename and stores it in the same folder as the source json files.}

\item{template}{The path to the .qmd file with the profile. If set to \code{NA}(default), the package's built-in profile is used.}

\item{anonymous}{If \code{TRUE}, all candidate names are replaced by A, B, C, ...}

\item{clear_cache}{The computations (e.g., citation statistics) are cached in the file \code{applicant_data.RData} (stored in the same folder as the jsons). If \code{clear_cache = TRUE}, the file is deleted and recomputed.}
}
\value{
The path to the rendered file
}
\description{
Render an HTML overview of multiple applicants
}
