% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_profile.R
\name{render_profile}
\alias{render_profile}
\title{Render an HTML or PDF profile of a single applicant}
\usage{
render_profile(json_path, output_file = NA)
}
\arguments{
\item{json_path}{The path to the applicant's JSON file}

\item{output_file}{The file name (optionally including a path) of the output report. If NA, it uses the last name from the applicant plus the current date-time as filename.}
}
\value{
The path to the rendered file
}
\description{
Render an HTML or PDF profile of a single applicant
}
