% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_profile.R
\name{render_profile}
\alias{render_profile}
\title{Render an HTML or PDF profile of a single applicant}
\usage{
render_profile(json_path, show_inter = TRUE, output_file = NA, template = NA)
}
\arguments{
\item{json_path}{The path to the applicant's JSON file}

\item{show_inter}{Show/hide the section on internationality and interdisciplinarity}

\item{output_file}{The file name (optionally including a path) of the output report. If NA, it uses the last name from the applicant plus the current date-time as filename.}

\item{template}{The path to the .Rmd file with the profile. If set to \code{NA}(default), the package's built-in profile is used.}
}
\value{
The path to the rendered file
}
\description{
Render an HTML or PDF profile of a single applicant
}
