% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRS_radarchart.R
\name{RRS_radarchart}
\alias{RRS_radarchart}
\title{Radar Chart for Relative Rigor Score}
\usage{
RRS_radarchart(
  RRS,
  overall_score = FALSE,
  minimal = FALSE,
  show.legend = FALSE,
  show.n_pub = TRUE,
  show.years = TRUE,
  base_size = 14,
  scale = "area"
)
}
\arguments{
\item{RRS}{The output from the \code{compute_RRS} function.}

\item{overall_score}{Show the overall score in the plot?}

\item{minimal}{If TRUE, a very simple chart without any legends is drawn}

\item{show.legend}{Show the legend to the right?}

\item{show.n_pub}{Show the number of publication on which the chart is based?}

\item{show.years}{Show a mini histogram to the right with the publication years?}

\item{base_size}{Base font size}

\item{scale}{Should the sectors be filled in a \code{"linear"} way (i.e., 50\% is 50\% of the radial height), or to the \code{"area"} (i.e., 50\% is 50\% of the area). Default = \code{"area"}.}
}
\value{
A radar chart visualizing the rigor profile for the applicant, showing the distribution of scores across different dimensions.
}
\description{
This function generates a radar chart that visualizes the rigor profile for a given applicant's data.
}
\examples{
radar_dat <- data.frame(
  dimension = c("Open Data", "Open Materials", "Preregistration",
    "Reproducible Code \n& Verification", "Theorizing \n& Formal Modeling"),
  max_points = rep(10, 5),
  rel_score=c(1, 0.75, 0.5, 0.25, 0.1),
  xstart = 0:4,
  xend = 1:5,
  xmid=0:4 + 0.5
)
RRS <- list(radar_dat=radar_dat, overall_score=0.5, n_papers=4, publication_years=c(2004, 2007, 2015, 2015))
RRS_radarchart(RRS)
RRS_radarchart(RRS, scale="linear")

}
