% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Score a single research output}
\usage{
score(research_output, verbose = FALSE)
}
\arguments{
\item{research_output}{The research output to be scored}

\item{verbose}{A logical value indicating whether to print verbose output (default is \code{FALSE})}
}
\value{
A list with the scoring information
}
\description{
Returns a list with the following elements:
\itemize{
\item \code{max_score}: the maximum score that can be reached
\item \code{score}: the score that was reached
\item \code{relative_score}: the score that was reached divided by the maximum score
}
}
